feat: enhanced daily summary, WhatsApp groups, fractional qty fixes, and README overhaul
Backend:
- Add daily_summary_service.py with comprehensive trading reports
  - Account P/L, GTT orders by symbol, filled today, locked BP
  - Replaces simple failed orders summary in gtt_service.py
- WhatsApp service now supports group notifications (WHATSAPP_GROUP_ID)
- Fix schemas.py: quantity field now float for fractional crypto support
- background_tasks.py: use new enhanced daily summary service

Frontend:
- GTTOrderCard: display fractional quantities correctly with formatQuantity
- GTTOrderCard: better error handling for Pydantic validation errors
- VersionUpdateDialog: show current vs new version in update prompt
- useGTTOrderActions: improved error message extraction
- useVersionCheck: correctly set current/latest version labels

Scripts:
- 50-setup-waha.sh: use shared WAHA at ~/Documents/waha
- 51-setup-waha-session.sh: better auth error handling, check docker logs first
- 52-get-waha-qr.sh: same API key extraction improvements
- 55-trigger-daily-summary.py: CLI wrapper with --dry-run option

Documentation:
- Complete README overhaul - concise, utilitarian, up-to-date
- Removed personal phone numbers from examples
- Added architecture section and tech stack details

+++ backend/schemas.py
@@ -66,7 +66,7 @@ class GTTOrderDetailUpdate(BaseModel):
     """Schema for updating an order detail."""
 
     trigger_price: float | None = None
-    quantity: int | None = None
+    quantity: float | None = None  # Changed to float to support fractional quantities (crypto)
     limit_price: float | None = None
     time_in_force: str | None = None
